+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 * [Polymer] Extract 'listeners' from 1.0-style declarations.
+* [Polymer] Extract pseudo elements from HTML comments
 
 ### Fixed
 * Fixed a class of race conditions and cache invalidation errors that can occur when there are concurrent analysis runs and edits to files.